# Mod03/DemoFiles/06_AutoRest/AutoRest/AutoRest.Host/Models/Destination.cs
namespace AutoRest.Host.Models
{
    public class Destination
    {
        public int Id { get; set; }

        public string CityName { get; set; }

        public string Airport { get; set; }
    }
}