# Mod03/DemoFiles/06_AutoRest/Starter/AutoRest.Sdk/Generated/Models/Destination.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace Autorest.Sdk.Models
{
    using Newtonsoft.Json;
    using System.Linq;

    public partial class Destination
    {
        /// <summary>
        /// Initializes a new instance of the Destination class.
        /// </summary>
        public Destination()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the Destination class.
        /// </summary>
        public Destination(int? id = default(int?), string cityName = default(string), string airport = default(string))
        {
            Id = id;
            CityName = cityName;
            Airport = airport;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "id")]
        public int? Id { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "cityName")]
        public string CityName { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "airport")]
        public string Airport { get; set; }

    }
}